Required lv3

+++ src/main/java/com/sparta/api/member/service/impl/MemberServiceImpl.java
@@ -1,5 +1,6 @@
 package com.sparta.api.member.service.impl;
 
+import com.sparta.api.member.dto.MemberDelDto;
 import com.sparta.api.member.dto.MemberModDto;
 import com.sparta.api.member.dto.MemberReqDto;
 import com.sparta.api.member.dto.MemberResDto;
@@ -30,7 +31,7 @@ public class MemberServiceImpl implements MemberService {
             throw new CustomException(CommonExceptionResultMessage.DUPLICATE_FAIL, "이미 사용 중인 이메일입니다.");
         }
 
-        Member member = new Member(dto.getName(), email);
+        Member member = new Member(dto.getName(), email, dto.getPassword());
         memberRepository.save(member);
 
         if (member.getId() == null) {
@@ -48,14 +49,20 @@ public class MemberServiceImpl implements MemberService {
     @Override
     public MemberResDto updateMember(Long id, MemberModDto dto) {
         Member member = this.getMember(id);
+        if (!member.getPassword().equals(dto.getPassword())) {
+            throw new CustomException(CommonExceptionResultMessage.PW_MISMATCH);
+        }
         member.update(dto.getName());
         memberRepository.save(member);
         return new MemberResDto(member);
     }
 
     @Override
-    public void deleteMember(Long id) {
+    public void deleteMember(Long id, MemberDelDto dto) {
         Member member = this.getMember(id);
+        if (!member.getPassword().equals(dto.getPassword())) {
+            throw new CustomException(CommonExceptionResultMessage.PW_MISMATCH);
+        }
         memberRepository.delete(member);
     }
